Finished working on the Room database

+++ android application/app/src/main/java/hogent/be/lunchers/models/ReservatieRepository.kt
@@ -6,10 +6,8 @@ class ReservatieRepository(private val orderDao: OrderDao) {
     val orders = orderDao.getAllOrders()
 
     fun insert(orders: List<Reservatie>) {
-        orderDao.insert(orders)
+        if (orders.isNotEmpty()) orderDao.deleteAllOrders()
+        orders.forEach { orderDao.insert(it) }
     }
 
-    fun delete() {
-        orderDao.deleteAllOrders()
-    }
 }
